# Testing/junk/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "The Luno API provides developers with a wealth of financial information provided through the Luno platform.  Through this secure system developers can:\n\n<ul>\n   <li>Create accounts for trading in cryptocurrencies</li>\n   <li>Access current and historic cryptocurrency market data</li>\n   <li>Submit trade orders and view order status</li>\n   <li>Buy and sell Bitcoin and Ethereum</li>\n   <li>Send and receive Bitcoin and Ethereum</li>\n   <li>Generate Bitcoin and Ethereum wallet addresses</li>\n </ul>\n\n The Luno API brings the world of Bitcoin and Ethereum to your doorstep.\n",
    "title": "Luno API",
    "version": "1.2.5"
  },
  "paths": {
...
    "/api/1/balance": {
      "get": {
        "description": "The list of all Accounts and their respective balances for the requesting user.\n\nPermissions required: <code>Perm_R_Balance</code>",
        "tags": [
          "Accounts"
        ],
        "summary": "List account balances",
        "operationId": "getBalances",
        "parameters": [
          {
            "example": "XBT",
            "x-go-name": "Assets",
            "description": "Only return balances for wallets with these currencies (if not provided,\nall balances will be returned). To request balances for multiple currencies,\npass the parameter multiple times,\ne.g. `assets=XBT&assets=ETH`.",
            "name": "assets",
            "in": "query",
            "style": "form",
            "explode": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getBalancesResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/1/fee_info": {
      "get": {
        "description": "Returns the fees and 30 day trading volume (as of midnight) for a given currency pair.  For complete details, please see <a href=\"en/countries\">Fees & Features</a>.\n\nPermissions required: <code>Perm_R_Orders</code>",
        "tags": [
          "Orders"
        ],
        "summary": "Get fee information",
        "operationId": "getFeeInfo",
        "parameters": [
          {
            "example": "XBTZAR",
            "x-go-name": "Pair",
            "description": "Get fee information about this pair.",
            "name": "pair",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getFeeInfoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/1/ticker": {
      "get": {
        "description": "Returns the latest ticker indicators for the specified currency pair.\n\nPlease see the <a href=\"#tag/currency \">Currency list</a> for the complete list of supported currency pairs.",
        "tags": [
          "Market"
        ],
        "summary": "Get ticker for currency pair",
        "operationId": "GetTicker",
        "parameters": [
          {
            "example": "XBTZAR",
            "x-go-name": "Pair",
            "description": "Currency pair",
            "name": "pair",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTickerResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/apiError"
          }
        }
      }
    },
    "/api/1/tickers": {
      "get": {
        "description": "Returns the latest ticker indicators from all active Luno exchanges.\n\nPlease see the <a href=\"#tag/currency \">Currency list</a> for the complete list of supported currency pairs.",
        "tags": [
          "Market"
        ],
        "summary": "List tickers for all currency pairs",
        "operationId": "GetTickers",
        "parameters": [
          {
            "example": "XBTZAR",
            "x-go-name": "Pair",
            "description": "Return tickers for multiple markets (if not provided, all tickers will be returned).\nTo request tickers for multiple markets, pass the parameter multiple times,\ne.g. `pair=XBTZAR&pair=ETHZAR`.",
            "name": "pair",
            "in": "query",
            "style": "form",
            "explode": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTickersResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/apiError"
          }
        }
      }
    },
...
    "/api/exchange/1/markets": {
      "get": {
        "description": "List all supported markets parameter information like price scale, min and\nmax order volumes and market ID.",
        "tags": [
          "Market"
        ],
        "summary": "Get markets info",
        "operationId": "Markets",
        "parameters": [
          {
            "example": "XBTZAR",
            "x-go-name": "Markets",
            "description": "List of market pairs to return. Requesting only the required pairs will improve response times.",
            "name": "pair",
            "in": "query",
            "style": "form",
            "explode": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketsInfoResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/apiError"
          }
        }
      }
    },
...
  },
...
  "servers": [
    {
      "url": "https://api.luno.com"
    }
  ],
  "components": {
    "responses": {
      "apiError": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "description": "Unique error reference",
                  "type": "string",
                  "x-go-name": "Code"
                },
                "message": {
                  "description": "Verbose error message",
                  "type": "string",
                  "x-go-name": "Message"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "AccountBalance": {
        "type": "object",
        "title": "Balance represents the current state of the account.",
        "properties": {
          "account_id": {
            "description": "ID of the account.",
            "type": "string",
            "format": "int64",
            "x-go-name": "AccountId",
            "example": "\"237592692\""
          },
          "asset": {
            "description": "Currency code for the asset held in this account.",
            "type": "string",
            "x-go-name": "Asset",
            "example": "\"XBT\""
          },
          "balance": {
            "description": "The amount available to send or trade.",
            "type": "string",
            "format": "amount",
            "x-go-name": "Balance"
          },
          "name": {
            "description": "The name set by the user upon creating the account.",
            "type": "string",
            "x-go-name": "Name",
            "example": "\"Trading account\""
          },
          "reserved": {
            "description": "Amount locked by Luno and cannot be sent or traded. This could be due to\nopen orders.",
            "type": "string",
            "format": "amount",
            "x-go-name": "Reserved"
          },
          "unconfirmed": {
            "description": "Amount that is awaiting some sort of verification to be credited to this\naccount. This could be an on-chain transaction that Luno is waiting for\nfurther block verifications to happen.",
            "type": "string",
            "format": "amount",
            "x-go-name": "Unconfirmed"
          }
        },
        "x-go-name": "Balance",
        "x-go-package": "bitx/services/exchange-api/internal/ops/publicapi"
      },
...
      "GetTickerResponse": {
        "description": "GetTickerResponse response for /api/ticker",
        "type": "object",
        "properties": {
          "ask": {
            "description": "The lowest ask price",
            "type": "string",
            "format": "amount",
            "x-go-name": "Ask"
          },
          "bid": {
            "description": "The highest bid price",
            "type": "string",
            "format": "amount",
            "x-go-name": "Bid"
          },
          "last_trade": {
            "description": "Last trade price",
            "type": "string",
            "format": "amount",
            "x-go-name": "Last"
          },
          "pair": {
            "type": "string",
            "x-go-name": "Pair"
          },
          "rolling_24_hour_volume": {
            "description": "24h rolling trade volume",
            "type": "string",
            "format": "amount",
            "x-go-name": "Rolling24HourVolume"
          },
          "status": {
            "description": "Market current status\n\n<code>ACTIVE</code> when the market is trading normally\n\n<code>POSTONLY</code> when the market has been suspended and only post-only orders will be accepted\n\n<code>DISABLED</code> when the market is shutdown and no orders can be accepted\n\n<code>UNKNOWN</code> the market status could not be determined. This is a temporary state.",
            "type": "string",
            "enum": [
              "ACTIVE",
              "POSTONLY",
              "DISABLED",
              "UNKNOWN"
            ],
            "x-go-name": "Status",
            "example": "ACTIVE"
          },
          "timestamp": {
            "description": "Unix timestamp in milliseconds of the tick",
            "type": "integer",
            "format": "timestamp",
            "x-go-name": "Timestamp"
          }
        },
        "x-go-package": "bitx/services/exchange-api/internal/ops/publicapi"
      },
...
      "ListTickersResponse": {
        "description": "ListTickersResponse response for /api/tickers",
        "type": "object",
        "properties": {
          "tickers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ticker"
            },
            "x-go-name": "Tickers"
          }
        },
        "x-go-package": "bitx/services/exchange-api/internal/ops/publicapi"
      },
...
      "MarketInfo": {
        "description": "MarketInfo current market information",
        "type": "object",
        "properties": {
          "base_currency": {
            "description": "Base currency code",
            "type": "string",
            "x-go-name": "BaseCcy",
            "example": "XBT"
          },
          "counter_currency": {
            "description": "Counter currency code",
            "type": "string",
            "x-go-name": "CounterCcy",
            "example": "EUR"
          },
          "fee_scale": {
            "description": "Fee decimal places",
            "type": "integer",
            "format": "int64",
            "x-go-name": "FeeScale"
          },
          "market_id": {
            "description": "Unique identifier for the market",
            "type": "string",
            "x-go-name": "MarketID",
            "example": "XBTEUR"
          },
          "max_price": {
            "description": "Maximum order price",
            "type": "string",
            "format": "amount",
            "x-go-name": "MaxPrice",
            "example": "100000.00"
          },
          "max_volume": {
            "description": "Maximum order volume",
            "type": "string",
            "format": "amount",
            "x-go-name": "MaxVolume",
            "example": "100.0"
          },
          "min_price": {
            "description": "Minimum order price",
            "type": "string",
            "format": "amount",
            "x-go-name": "MinPrice",
            "example": "100.00"
          },
          "min_volume": {
            "description": "Minimum order volume",
            "type": "string",
            "format": "amount",
            "x-go-name": "MinVolume",
            "example": "0.0005"
          },
          "price_scale": {
            "description": "Price decimal places",
            "type": "integer",
            "format": "int64",
            "x-go-name": "PriceScale",
            "example": 2
          },
          "trading_status": {
            "description": "Current market trading status:<br>\n<code>POST_ONLY</code> Trading is indefinitely suspended. This state is\ncommonly used when new markets are being launched to give traders enough\ntime to setup their orders before trading begins. When in this status,\norders can only be posted as post-only.<br>\n<code>ACTIVE</code> Trading is fully enabled.<br>\n<code>SUSPENDED</code> Trading has been temporarily suspended due to very\nhigh volatility. When in this status, orders can only be posted as\npost-only.<br>\n<code>Unknown</code> Trading status is unknown. This could indicate a temporary error\non the market and should resolve shortly.",
            "type": "string",
            "enum": [
              "POST_ONLY",
              "ACTIVE",
              "SUSPENDED",
              "UNKNOWN"
            ],
            "x-go-name": "TradingStatus"
          },
          "volume_scale": {
            "description": "Volume decimal places",
            "type": "integer",
            "format": "int64",
            "x-go-name": "VolumeScale",
            "example": 4
          }
        },
        "x-go-package": "bitx/services/exchange-api/internal/ops/publicapi"
      },
      "MarketsInfoResponse": {
        "description": "MarketsInfoResponse response for /api/2/markets",
        "type": "object",
        "properties": {
          "markets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketInfo"
            },
            "x-go-name": "Markets"
          }
        },
        "x-go-package": "bitx/services/exchange-api/internal/ops/publicapi"
      },
...
      "Ticker": {
        "description": "Ticker summarised instant market data",
        "type": "object",
        "properties": {
          "ask": {
            "description": "The lowest ask price",
            "type": "string",
            "format": "amount",
            "x-go-name": "Ask"
          },
          "bid": {
            "description": "The highest bid price",
            "type": "string",
            "format": "amount",
            "x-go-name": "Bid"
          },
          "last_trade": {
            "description": "Last trade price",
            "type": "string",
            "format": "amount",
            "x-go-name": "Last"
          },
          "pair": {
            "type": "string",
            "x-go-name": "Pair"
          },
          "rolling_24_hour_volume": {
            "description": "24h rolling trade volume",
            "type": "string",
            "format": "amount",
            "x-go-name": "Rolling24HourVolume"
          },
          "status": {
            "description": "Market current status\n\n<code>ACTIVE</code> when the market is trading normally\n\n<code>POSTONLY</code> when the market has been suspended and only post-only orders will be accepted\n\n<code>DISABLED</code> when the market is shutdown and no orders can be accepted\n\n<code>UNKNOWN</code> the market status could not be determined. This is a temporary state.",
            "type": "string",
            "enum": [
              "ACTIVE",
              "POSTONLY",
              "DISABLED",
              "UNKNOWN"
            ],
            "x-go-name": "Status",
            "example": "ACTIVE"
          },
          "timestamp": {
            "description": "Unix timestamp in milliseconds of the tick",
            "type": "integer",
            "format": "timestamp",
            "x-go-name": "Timestamp"
          }
        },
        "x-go-package": "bitx/services/exchange-api/internal/ops/publicapi"
      },
...
      "getBalancesResponse": {
        "description": "GetBalancesResponse balance lookup response",
        "type": "object",
        "properties": {
          "balance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountBalance"
            },
            "x-go-name": "Balance"
          }
        },
        "x-go-name": "GetBalancesResponse",
        "x-go-package": "bitx/services/exchange-api/internal/ops/publicapi"
      },
      "getFeeInfoResponse": {
        "description": "GetFeeInfoResponse response object for get fee info calls",
        "type": "object",
        "properties": {
          "maker_fee": {
            "type": "string",
            "x-go-name": "MakerFee"
          },
          "taker_fee": {
            "type": "string",
            "x-go-name": "TakerFee"
          },
          "thirty_day_volume": {
            "type": "string",
            "x-go-name": "ThirtyDayVolume"
          }
        },
        "x-go-name": "GetFeeInfoResponse",
        "x-go-package": "bitx/services/exchange-api/internal/ops/publicapi"
      },
...
    }
  }
}